requirements.txt
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
numpy==1.16.2
opencv-contrib-python==4.1.0.25
Pillow==5.4.1